Agent: Move get_credentials() to credentials.py
