+++ monkey/infection_monkey/exploit/powershell.py
@@ -19,8 +19,7 @@ from infection_monkey.exploit.powershell_utils.auth_options import (
     AuthOptions,
     get_auth_options,
 )
-from infection_monkey.exploit.powershell_utils.credential_generators import get_credentials
-from infection_monkey.exploit.powershell_utils.credentials import Credentials
+from infection_monkey.exploit.powershell_utils.credentials import Credentials, get_credentials
 from infection_monkey.exploit.powershell_utils.utils import (
     IClient,
     get_client_based_on_auth_options,
