+++ monkey/tests/unit_tests/infection_monkey/exploit/powershell_utils/test_credentials.py
@@ -1,5 +1,4 @@
-from infection_monkey.exploit.powershell_utils.credential_generators import get_credentials
-from infection_monkey.exploit.powershell_utils.credentials import Credentials
+from infection_monkey.exploit.powershell_utils.credentials import Credentials, get_credentials
 
 TEST_USERNAMES = ["user1", "user2"]
 TEST_PASSWORDS = ["p1", "p2"]
